Moved docs to regx.github.com
RegX updates will be rapidly synced to this account.  Once a feature is
ready to be rolled out, it'll be pushed to the regx.github.com account
with documentation.  If you want to use RegX to play around with in your
scripts, be sure to use the file here: regx.github.com/RegX.js